fixed Clock class and renamed some classes
